feat(t8s-cluster): Allow for parallel image pulling with k8s v1.27
This allows to configure a max number of images pulled in parallel and turns
off serializeImagePulls in that case. This only works with k8s v1.27 and up.
The default is keept to 1 for now to test the functionality on specific
clusters over a periode of time. It will probabbly be raised later on.

References:
- #418